progress and clean
