+++ tests/envs/test_pybullet_robots.py
@@ -11,7 +11,8 @@ from predicators.src.pybullet_helpers.robots.panda import PandaPyBulletRobot
 from predicators.src.pybullet_helpers.robots.single_arm import \
     create_single_arm_pybullet_robot
 from predicators.src.pybullet_helpers.utils import get_kinematic_chain, \
-    pybullet_inverse_kinematics, run_motion_planning
+    pybullet_inverse_kinematics
+from predicators.src.pybullet_helpers.motion_planning import run_motion_planning
 from predicators.src.settings import CFG
